+++ lib/jxl/quant_weights_test.cc
@@ -232,8 +232,9 @@ TEST_P(QuantWeightsTargetTest, DCTUniform) {
 
     for (size_t i = 0; i < 64; i++) {
       // DCTSlow doesn't multiply/divide by 1/N, so we do it manually.
-      slow_coeffs[i] = roundf(slow_coeffs[i] / kUniformQuant) * kUniformQuant;
-      coeffs[i] = roundf(coeffs[i] / dequant_matrices->Matrix(dct, 0)[i]) *
+      slow_coeffs[i] =
+          std::round(slow_coeffs[i] / kUniformQuant) * kUniformQuant;
+      coeffs[i] = std::round(coeffs[i] / dequant_matrices->Matrix(dct, 0)[i]) *
                   dequant_matrices->Matrix(dct, 0)[i];
     }
     IDCTSlow<8>(slow_coeffs);
@@ -252,8 +253,9 @@ TEST_P(QuantWeightsTargetTest, DCTUniform) {
     DCTSlow<16>(slow_coeffs);
 
     for (size_t i = 0; i < 64 * 4; i++) {
-      slow_coeffs[i] = roundf(slow_coeffs[i] / kUniformQuant) * kUniformQuant;
-      coeffs[i] = roundf(coeffs[i] / dequant_matrices->Matrix(dct, 0)[i]) *
+      slow_coeffs[i] =
+          std::round(slow_coeffs[i] / kUniformQuant) * kUniformQuant;
+      coeffs[i] = std::round(coeffs[i] / dequant_matrices->Matrix(dct, 0)[i]) *
                   dequant_matrices->Matrix(dct, 0)[i];
     }
